+++ pg/src/driver/core.rs
@@ -166,6 +166,7 @@ impl PgDriver {
             gss_token_provider: pool_cfg.gss_token_provider,
             gss_token_provider_ex: pool_cfg.gss_token_provider_ex,
             auth: pool_cfg.auth_settings,
+            io_uring: pool_cfg.io_uring,
             startup_params: Vec::new(),
         };